Use pagination for big arrays always
